flight-operationのflightplanのアダプター構成をリファクタリング
