+++ flight-operation/pkg/flightplan/adapters/rabbitmq/mock_data_test.go
@@ -47,11 +47,28 @@ func (ch *channelMockPublish) Close() error {
 	panic("implement me")
 }
 
-type publishHandlerMock struct {
+type pubSubManagerMock struct {
+	consumers       []consumer
 	publishHandlers []func(ch crm.Channel, e interface{})
 }
 
-func (h *publishHandlerMock) SetPublishHandler(handler func(ch crm.Channel, e interface{})) error {
+func (h *pubSubManagerMock) SetConsumer(ctx context.Context, exchangeName, queueName string, handler func([]byte)) error {
+	h.consumers = append(
+		h.consumers,
+		consumer{
+			exchangeName: exchangeName,
+			queueName:    queueName,
+			handler:      handler,
+		})
+	return nil
+}
+
+func (h *pubSubManagerMock) SetPublishHandler(handler func(ch crm.Channel, e interface{})) error {
 	h.publishHandlers = append(h.publishHandlers, handler)
 	return nil
+}
+
+type consumer struct {
+	exchangeName, queueName string
+	handler                 func([]byte)
 }
