+++ flight-operation/pkg/flightplan/adapters/rabbitmq/subscribe_event_publisher.go
@@ -8,8 +8,8 @@ import (
 	"github.com/golang/glog"
 )
 
-// SubscribePublishHandler .
-func SubscribePublishHandler(psm crm.PubSubManagerSetter) {
+// SubscribeEventPublisher .
+func SubscribeEventPublisher(psm crm.PubSubManagerSetter) {
 	psm.SetPublishHandler(
 		func(ch crm.Channel, e interface{}) {
 			if event, ok := e.(flightplan.FleetIDGaveEvent); ok {
